requirements.txt
# Telegram Bot Library
python-telegram-bot[job-queue]

# Supabase Python Client
supabase

# Google Gemini AI Client
google-generativeai

# For loading .env files
python-dotenv
